+++ changelog.md
@@ -7,7 +7,7 @@
  - added password encryption for Login [password-encryption]
  - classes list coming in descending order [Class-Serialised]
   - Subjects list coming in descending order [Subject-Serialised]
- - classes list coming in descending order [Class-Serialised]
+ - added BreadCrumbs in Class , subject and User Module [BreadCrumbs]
 
  
  # Fixed